feat(base test):Rename base test to refer like dislike funcionality
-Rename base_test file to like_dislike_base_test

[Fixes #162948845]
